streamlit/requirements.txt
streamlit==1.10.0 
requests-toolbelt==0.9.1
altair<5

fastapi==0.68.0
pillow==9.0.1
torch==1.12.1
torchvision==0.13.1
python-multipart==0.0.5 
aiofiles==0.4.0

# YOLOv5 requirements
# Base ------------------------------------------------------------------------
gitpython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
psutil  # system resources
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1  # FLOPs computation
tqdm>=4.64.0
ultralytics
# protobuf==3.20.1  # https://github.com/ultralytics/yolov5/issues/8012

# # Logging ---------------------------------------------------------------------
# tensorboard>=2.4.1

# # Plotting --------------------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0